requirements.txt
flask
structlog>=24,<26
python-dotenv
requests
pymongo>=4.0,<5
mongoengine>=0.27.0
pytz
gunicorn
raven[flask]
pypdf2>=3.0.0
